+++ projects/rocprofiler-sdk/source/lib/rocprofiler-sdk-tool/tool.cpp
@@ -215,11 +215,11 @@ struct buffer_ids
     rocprofiler_buffer_id_t pc_sampling_host_trap   = {};
     rocprofiler_buffer_id_t rocdecode_api_trace     = {};
     rocprofiler_buffer_id_t rocjpeg_api_trace       = {};
-    rocprofiler_buffer_id_t hipfile_api_trace       = {};
     rocprofiler_buffer_id_t pc_sampling_stochastic  = {};
     rocprofiler_buffer_id_t ompt_trace              = {};
     rocprofiler_buffer_id_t hip_graph_trace         = {};
     rocprofiler_buffer_id_t rocshmem_api_trace      = {};
+    rocprofiler_buffer_id_t hipfile_api_trace       = {};
 
     auto as_array() const
     {
@@ -235,11 +235,11 @@ struct buffer_ids
                                                        pc_sampling_host_trap,
                                                        rocdecode_api_trace,
                                                        rocjpeg_api_trace,
-                                                       hipfile_api_trace,
                                                        pc_sampling_stochastic,
                                                        ompt_trace,
                                                        hip_graph_trace,
-                                                       rocshmem_api_trace};
+                                                       rocshmem_api_trace,
+                                                       hipfile_api_trace};
     }
     auto pc_sampling_buffers_as_array() const
     {
@@ -3685,12 +3685,12 @@ generate_output(cleanup_mode _cleanup_mode, bool skip_output = false)
                          memory_allocation_output.get_generator(),
                          rocdecode_output.get_generator(),
                          rocjpeg_output.get_generator(),
-                         hipfile_output.get_generator(),
                          pc_sampling_host_trap_output.get_generator(),
                          pc_sampling_stochastic_output.get_generator(),
                          spm_counters_output.get_generator(),
                          hip_graph_output.get_generator(),
-                         rocshmem_output.get_generator());
+                         rocshmem_output.get_generator(),
+                         hipfile_output.get_generator());
         json_ar.finish_process();
 
         tool::close_json(json_ar);
@@ -3731,12 +3731,12 @@ generate_output(cleanup_mode _cleanup_mode, bool skip_output = false)
                           kfd_output.get_generator(),
                           rccl_output.get_generator(),
                           rocdecode_output.get_generator(),
-                          hipfile_output.get_generator(),
                           counters_output.get_generator(),
                           spm_counters_output.get_generator(),
                           ompt_output.get_generator(),
                           hip_graph_output.get_generator(),
-                          rocshmem_output.get_generator());
+                          rocshmem_output.get_generator(),
+                          hipfile_output.get_generator());
     }
 
     if(tool::get_config().otf2_output && outdata.num_output > 0 &&