+++ projects/rocprofiler-sdk/source/lib/rocprofiler-sdk/buffer_tracing.cpp
@@ -103,8 +103,6 @@ ROCPROFILER_BUFFER_TRACING_KIND_STRING(OMPT)
 ROCPROFILER_BUFFER_TRACING_KIND_STRING(RUNTIME_INITIALIZATION)
 ROCPROFILER_BUFFER_TRACING_KIND_STRING(ROCDECODE_API)
 ROCPROFILER_BUFFER_TRACING_KIND_STRING(ROCJPEG_API)
-ROCPROFILER_BUFFER_TRACING_KIND_STRING(HIPFILE_API)
-ROCPROFILER_BUFFER_TRACING_KIND_STRING(HIPFILE_API_EXT)
 ROCPROFILER_BUFFER_TRACING_KIND_STRING(HIP_STREAM)
 ROCPROFILER_BUFFER_TRACING_KIND_STRING(HIP_RUNTIME_API_EXT)
 ROCPROFILER_BUFFER_TRACING_KIND_STRING(HIP_COMPILER_API_EXT)
@@ -121,6 +119,8 @@ ROCPROFILER_BUFFER_TRACING_KIND_STRING(MARKER_CORE_RANGE_API)
 ROCPROFILER_BUFFER_TRACING_KIND_STRING(HIP_GRAPH)
 ROCPROFILER_BUFFER_TRACING_KIND_STRING(ROCSHMEM_API)
 ROCPROFILER_BUFFER_TRACING_KIND_STRING(ROCSHMEM_API_EXT)
+ROCPROFILER_BUFFER_TRACING_KIND_STRING(HIPFILE_API)
+ROCPROFILER_BUFFER_TRACING_KIND_STRING(HIPFILE_API_EXT)
 
 template <size_t Idx, size_t... Tail>
 std::pair<const char*, size_t>
@@ -339,12 +339,6 @@ rocprofiler_query_buffer_tracing_kind_operation_name(rocprofiler_buffer_tracing_
             val = rocprofiler::rocjpeg::name_by_id<ROCPROFILER_ROCJPEG_TABLE_ID_CORE>(operation);
             break;
         }
-        case ROCPROFILER_BUFFER_TRACING_HIPFILE_API:
-        case ROCPROFILER_BUFFER_TRACING_HIPFILE_API_EXT:
-        {
-            val = rocprofiler::hipfile::name_by_id<ROCPROFILER_HIPFILE_TABLE_ID_CORE>(operation);
-            break;
-        }
         case ROCPROFILER_BUFFER_TRACING_HIP_STREAM:
         {
             val = rocprofiler::hip::stream::name_by_id(operation);
@@ -377,6 +371,12 @@ rocprofiler_query_buffer_tracing_kind_operation_name(rocprofiler_buffer_tracing_
         case ROCPROFILER_BUFFER_TRACING_ROCSHMEM_API_EXT:
         {
             val = rocprofiler::rocshmem::name_by_id<ROCPROFILER_ROCSHMEM_TABLE_ID_CORE>(operation);
+            break;
+        }
+        case ROCPROFILER_BUFFER_TRACING_HIPFILE_API:
+        case ROCPROFILER_BUFFER_TRACING_HIPFILE_API_EXT:
+        {
+            val = rocprofiler::hipfile::name_by_id<ROCPROFILER_HIPFILE_TABLE_ID_CORE>(operation);
             break;
         }
     };
@@ -518,12 +518,6 @@ rocprofiler_iterate_buffer_tracing_kind_operations(
             ops = rocprofiler::rocjpeg::get_ids<ROCPROFILER_ROCJPEG_TABLE_ID_CORE>();
             break;
         }
-        case ROCPROFILER_BUFFER_TRACING_HIPFILE_API:
-        case ROCPROFILER_BUFFER_TRACING_HIPFILE_API_EXT:
-        {
-            ops = rocprofiler::hipfile::get_ids<ROCPROFILER_HIPFILE_TABLE_ID_CORE>();
-            break;
-        }
         case ROCPROFILER_BUFFER_TRACING_HIP_STREAM:
         {
             ops = rocprofiler::hip::stream::get_ids();
@@ -555,6 +549,12 @@ rocprofiler_iterate_buffer_tracing_kind_operations(
         case ROCPROFILER_BUFFER_TRACING_ROCSHMEM_API_EXT:
         {
             ops = rocprofiler::rocshmem::get_ids<ROCPROFILER_ROCSHMEM_TABLE_ID_CORE>();
+            break;
+        }
+        case ROCPROFILER_BUFFER_TRACING_HIPFILE_API:
+        case ROCPROFILER_BUFFER_TRACING_HIPFILE_API_EXT:
+        {
+            ops = rocprofiler::hipfile::get_ids<ROCPROFILER_HIPFILE_TABLE_ID_CORE>();
             break;
         }
     }
@@ -592,6 +592,7 @@ rocprofiler_iterate_buffer_tracing_record_args(
         case ROCPROFILER_BUFFER_TRACING_MEMORY_COPY:
         case ROCPROFILER_BUFFER_TRACING_RCCL_API:
         case ROCPROFILER_BUFFER_TRACING_ROCSHMEM_API:
+        case ROCPROFILER_BUFFER_TRACING_HIPFILE_API:
         {
             return ROCPROFILER_STATUS_ERROR_NOT_IMPLEMENTED;
         }